+++ helpers/utils.h
@@ -2,6 +2,6 @@
 
 namespace Utils {
   unsigned int get_random_pos(unsigned int max_range, unsigned int factor);
-  unsigned int random(unsigned int start,  unsigned int end);
+  double random(int start, int end);
   bool passed_debounce_time(int last_tick);
 }
